Finished up Part 4 exercises

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,19 @@
 require('express-async-errors')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/Blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+
+/** 
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if(authorization && authorization.startsWith('bearer ')){
+    return authorization.replace('bearer ', '')
+  }
+  return null
+}
+*/
 
 blogsRouter.get('/', async (request, response) => {
   /** 
@@ -10,24 +23,39 @@ blogsRouter.get('/', async (request, response) => {
         response.json(blogs)
       })
       */
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
+    //console.log(blogs)
     response.json(blogs)
+    
+
   })
   
-  blogsRouter.post('/', async (request, response) => {
+  blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
-
+    //console.log(request.get('authorization'))
+    /** 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+      response.status(401).json({error: 'token invalid'})
+    }
     //console.log(body)
-
+    */
     if(!body.title || !body.url){
       response.status(400).end()
     }
     else{
+
+      //const user = await User.findById(body.userId)
+      //const user = await User.findById(decodedToken.id)
+      //const user = users[0]
+      const user = await User.findById(request.user)
+      //console.log(user)
       const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user.id
       })
       /** 
       blog
@@ -37,11 +65,30 @@ blogsRouter.get('/', async (request, response) => {
         })
         */
       const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
       response.status(201).json(savedBlog)
       }
   })
 
-  blogsRouter.delete('/:id', async (request, response) => {
+  blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+    /** 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+      response.status(401).json({error: 'token invalid'})
+    }
+    console.log(decodedToken)
+    */
+    console.log(request.params.id)
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+    //const user = await User.findById(decodedToken.id)
+    const user = await User.findById(request.user)
+    if(blog && (blog.user.toString() !== user._id.toString())){
+      return response.status(401).json({error: 'invalid user'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   })
